Refactoring: HomeScreenComposable.kt divided into BackLayerContent.kt, FrontLayerContent.kt, CurrencyRow.kt, SwapButton.kt files.
