+++ androidApp/src/main/java/com/tsymbaliuk/currency/converter/android/ui/home/CurrencyRow.kt
@@ -15,7 +15,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.TextFieldValue
@@ -27,75 +26,7 @@ import com.tsymbaliuk.currency.converter.android.ui.theme.*
 
 @Preview
 @Composable
-fun BackLayerContent(
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(AppColors.primary),
-        verticalArrangement = Arrangement.SpaceEvenly
-    ) {
-        Tittle(modifier = Modifier)
-        CurrencyRow(modifier = Modifier)
-    }
-}
-
-@Preview
-@Composable
-fun SwapButton(modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .padding(paddingMedium)
-            .width(120.dp)
-            .height(48.dp)
-            .clip(AppShapes.medium)
-            .background(AppColors.secondary)
-            .clickable { }
-    ) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_round_swap_vert_24),
-            contentDescription = stringResource(id = R.string.swap),
-            modifier = Modifier
-                .size(24.dp)
-                .align(Alignment.Center),
-            tint = Black
-        )
-    }
-}
-
-@Preview
-@Composable
-fun FrontLayerContent(
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .clip(AppShapes.large)
-            .background(GreyLight)
-    ) {
-        CurrencyRow(modifier = Modifier)
-    }
-
-}
-
-@Preview
-@Composable
-private fun Tittle(
-    modifier: Modifier = Modifier
-) {
-    Text(
-        modifier = modifier.padding(paddingMedium),
-        text = stringResource(id = R.string.currency_converter),
-        color = AppColors.secondary,
-        style = AppTypography.displaySmall
-    )
-}
-
-@Preview
-@Composable
-private fun CurrencyRow(
+fun CurrencyRow(
     modifier: Modifier = Modifier
 ) {
     Row(
